Refactor ResourcesCollectionProvider
