+++ src/Sylius/Bundle/ResourceBundle/Controller/ResourcesResolver.php
@@ -1,0 +1,43 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Bundle\ResourceBundle\Controller;
+
+use Sylius\Component\Resource\Repository\RepositoryInterface;
+
+/**
+ * @author Paweł Jędrzejewski <pawel@sylius.org>
+ */
+class ResourcesResolver implements ResourcesResolverInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getResources(RequestConfiguration $requestConfiguration, RepositoryInterface $repository)
+    {
+        if (null !== $repositoryMethod = $requestConfiguration->getRepositoryMethod()) {
+            $callable = [$repository, $repositoryMethod];
+            $resources = call_user_func_array($callable, $requestConfiguration->getRepositoryArguments());
+
+            return $resources;
+        }
+
+        if (!$requestConfiguration->isPaginated() && !$requestConfiguration->isLimited()) {
+            return $repository->findAll();
+        }
+
+        if (!$requestConfiguration->isPaginated()) {
+            return $repository->findBy($requestConfiguration->getCriteria(), $requestConfiguration->getSorting(), $requestConfiguration->getLimit());
+        }
+
+        return $repository->createPaginator($requestConfiguration->getCriteria(), $requestConfiguration->getSorting());
+    }
+}
