+++ src/Sylius/Bundle/ResourceBundle/Controller/ResourcesResolverInterface.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the Sylius package.
+ *
+ * (c) Paweł Jędrzejewski
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sylius\Bundle\ResourceBundle\Controller;
+
+use Sylius\Component\Resource\Repository\RepositoryInterface;
+
+/**
+ * @author Paweł Jędrzejewski <pawel@sylius.org>
+ */
+interface ResourcesResolverInterface
+{
+    /**
+     * @param RequestConfiguration $requestConfiguration
+     * @param RepositoryInterface $repository
+     *
+     * @return mixed
+     */
+    public function getResources(RequestConfiguration $requestConfiguration, RepositoryInterface $repository);
+}
